fix(vault): re-broadcast when an unlock lands on an already-unlocked vault
The unlock page renders nothing locally on `ok` — the state broadcast is what
unmounts it — and `broadcastToReplicas` swallows delivery failures. A dropped
broadcast was therefore unrecoverable: the spinner stopped, the screen did not
change, and resubmitting reached the already-unlocked path, which answers `ok`
without dispatching, so no further broadcast was ever produced. Only reopening
the popup helped, and nothing said so.

Broadcasting from that path makes the retry heal it.

`broadcastToReplicas` moves out of `get-main-store` into its own module with a
`broadcastPopupState` wrapper, so the store subscriber and this handler share
one definition instead of assembling the same event twice. The module is kept
light on purpose: importing `get-main-store` from a handler test would pull in
store creation.
